requirements.txt
streamlit
pandas
pyairtable
python-dotenv